# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PsychGen BR — AI-driven psychometric instrument development
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pipeline/jobs:
    get:
      operationId: listPipelineJobs
      tags: [pipeline]
      summary: List all pipeline jobs (optionally filter by project or stage)
      parameters:
        - name: projectId
          in: query
          required: false
          schema: { type: integer }
        - name: stage
          in: query
          required: false
          schema:
            type: string
            enum: [aigenie, difficulty, irt, sample_design]
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PipelineJob"
  /pipeline/jobs/{id}:
    get:
      operationId: getPipelineJob
      tags: [pipeline]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineJob"
        "404":
          $ref: "#/components/responses/NotFound"
  /pipeline/jobs/{id}/cancel:
    post:
      operationId: cancelPipelineJob
      tags: [pipeline]
      summary: Cancel a queued or running job
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineJob"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/runs/aigenie:
    post:
      operationId: runAigenieStage
      tags: [pipeline]
      summary: Stage 1 — generate items via AIGENIE-style LLM loop with EGA verification
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AigenieRunInput"
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineJob"
        "400":
          $ref: "#/components/responses/BadRequest"
  /projects/{id}/runs/difficulty:
    post:
      operationId: runDifficultyStage
      tags: [pipeline]
      summary: Stage 2 — train and apply ML difficulty prediction (glmnet + randomForest)
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DifficultyRunInput"
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineJob"
        "400":
          $ref: "#/components/responses/BadRequest"
  /projects/{id}/runs/irt:
    post:
      operationId: runIrtStage
      tags: [pipeline]
      summary: Stage 3 — IRT calibration via synthetic LLM respondents (mirt)
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IrtRunInput"
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineJob"
        "400":
          $ref: "#/components/responses/BadRequest"
  /projects/{id}/runs/sample-design:
    post:
      operationId: runSampleDesignStage
      tags: [pipeline]
      summary: Stage 5 — sample design (post-stratification weights + item shortlist)
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SampleDesignRunInput"
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineJob"
        "400":
          $ref: "#/components/responses/BadRequest"
  /projects/{id}/export.xlsx:
    get:
      operationId: exportProjectXlsx
      tags: [exports]
      summary: Export full project workbook (items + calibrations + Wright map data) as XLSX
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Excel workbook
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema: { type: string, format: binary }
        "404":
          $ref: "#/components/responses/NotFound"
  /pipeline/jobs/{id}/logs:
    get:
      operationId: getPipelineJobLogs
      tags: [pipeline]
      summary: Recent log + progress events for a job (REST snapshot of the SSE stream)
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Recent events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobLogEvent"
  /items/{id}:
    get:
      operationId: getItem
      tags: [items]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateItem
      tags: [items]
      summary: Edit item text or human-review status
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItemInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    AigenieParams:
      type: object
      description: Full AIGENIE parameter surface — every knob exposed to the user
      required:
        - model
        - temperature
        - topP
        - targetN
        - adaptive
        - allTogether
        - runOverall
        - embeddingModel
      properties:
        model: { type: string, default: "gpt-4o" }
        temperature: { type: number, minimum: 0, maximum: 2, default: 0.9 }
        topP: { type: number, minimum: 0, maximum: 1, default: 1.0 }
        targetN: { type: integer, minimum: 5, maximum: 200, default: 30 }
        adaptive: { type: boolean, default: true }
        allTogether: { type: boolean, default: false }
        runOverall: { type: boolean, default: true }
        systemRole:
          type: ["string", "null"]
          description: Custom system role for the LLM
        promptNotes:
          type: ["string", "null"]
          description: Extra notes appended to prompt
        itemAttributes:
          type: array
          items: { type: string }
          description: Constraints/attributes each item must satisfy
        itemExamples:
          type: array
          items: { type: string }
          description: Few-shot example items
        embeddingModel: { type: string, default: "text-embedding-3-large" }
        egaThreshold: { type: number, minimum: 0, maximum: 1, default: 0.7 }
    AigenieRunInput:
      type: object
      required: [params]
      properties:
        params: { $ref: "#/components/schemas/AigenieParams" }
    DifficultyParams:
      type: object
      properties:
        algorithm:
          type: string
          enum: [glmnet, randomForest, ensemble]
          default: ensemble
        crossValidationFolds: { type: integer, minimum: 2, maximum: 20, default: 5 }
        useTextFeatures: { type: boolean, default: true }
        useEmbeddingFeatures: { type: boolean, default: true }
        embeddingModel: { type: string, default: "text-embedding-3-large" }
    DifficultyRunInput:
      type: object
      required: [params]
      properties:
        params: { $ref: "#/components/schemas/DifficultyParams" }
    IrtParams:
      type: object
      description: IRT calibration via synthetic LLM respondents
      required: [models, syntheticN, irtModel]
      properties:
        models:
          type: array
          minItems: 1
          items:
            type: string
          description: LLM persona models in the ensemble (e.g. gpt-3.5-turbo, gpt-4o, claude-haiku)
          default: ["gpt-3.5-turbo", "gpt-4o-mini", "claude-haiku-4-5"]
        syntheticN: { type: integer, minimum: 50, maximum: 5000, default: 500 }
        irtModel:
          type: string
          enum: [Rasch, "2PL", "3PL", graded]
          default: "2PL"
        personaSeed:
          type: ["string", "null"]
          description: Persona generation system prompt seed
        temperature: { type: number, minimum: 0, maximum: 2, default: 1.0 }
        responseFormat:
          type: string
          enum: [likert5, likert7, dichotomous]
          default: likert5
    IrtRunInput:
      type: object
      required: [params]
      properties:
        params: { $ref: "#/components/schemas/IrtParams" }
    SampleDesignParams:
      type: object
      description: Stage 5 — sample design parameters (post-stratification + item info)
      required: [targetSampleN, strata, targetThetaSE]
      properties:
        targetSampleN: { type: integer, minimum: 30, maximum: 100000, default: 500 }
        targetThetaSE: { type: number, minimum: 0.05, maximum: 1.0, default: 0.3 }
        strata:
          type: array
          minItems: 1
          description: Population strata with target population shares (must sum to ~1)
          items:
            type: object
            required: [label, populationShare]
            properties:
              label: { type: string, minLength: 1 }
              populationShare: { type: number, minimum: 0, maximum: 1 }
              sampledN:
                type: ["integer", "null"]
                minimum: 0
                description: Optional already-collected N per stratum (for weight calc)
        shortlistMaxItems:
          type: ["integer", "null"]
          minimum: 5
          maximum: 500
          default: 30
          description: Number of items to keep in the high-information shortlist
    SampleDesignRunInput:
      type: object
      required: [params]
      properties:
        params: { $ref: "#/components/schemas/SampleDesignParams" }
    JobLogEvent:
      type: object
      required: [type, ts, message]
      properties:
        type: { type: string, enum: [progress, log] }
        ts: { type: string, format: date-time }
        message: { type: string }
        progress: { type: ["number", "null"], minimum: 0, maximum: 1 }
        level:
          type: ["string", "null"]
          enum: [info, warn, error, null]
    PipelineJob:
      type: object
      required:
        - id
        - projectId
        - stage
        - status
        - progress
        - paramsJson
        - createdAt
        - updatedAt
      properties:
        id: { type: integer }
        projectId: { type: integer }
        stage:
          type: string
          enum: [aigenie, difficulty, irt, sample_design]
        status:
          type: string
          enum: [queued, running, completed, failed, cancelled]
        progress: { type: number, minimum: 0, maximum: 1 }
        message:
          type: ["string", "null"]
        paramsJson:
          type: object
          additionalProperties: true
        resultJson:
          type: ["object", "null"]
          additionalProperties: true
        error:
          type: ["string", "null"]
        startedAt:
          type: ["string", "null"]
          format: date-time
        finishedAt:
          type: ["string", "null"]
          format: date-time
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Item:
      type: object
      required:
        - id
        - projectId
        - text
        - status
        - generatedBy
        - createdAt
        - updatedAt
      properties:
        id: { type: integer }
        projectId: { type: integer }
        text: { type: string }
        construct:
          type: ["string", "null"]
        dimension:
          type: ["string", "null"]
        status:
          type: string
          enum: [generated, needs_review, approved, rejected]
        generatedBy: { type: string, description: "Model that generated the item" }
        difficultyPredicted:
          type: ["number", "null"]
        difficultyEstimated:
          type: ["number", "null"]
        discrimination:
          type: ["number", "null"]
        guessing:
          type: ["number", "null"]
        egaCommunity:
          type: ["integer", "null"]
        humanNotes:
          type: ["string", "null"]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    UpdateItemInput:
      type: object
      properties:
        text: { type: string, minLength: 1 }
        status:
          type: string
          enum: [generated, needs_review, approved, rejected]
        humanNotes:
          type: ["string", "null"]
        construct:
          type: ["string", "null"]
        dimension:
          type: ["string", "null"]
...