# lib/api-zod/src/generated/types/difficultyParams.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PsychGen BR — AI-driven psychometric instrument development
 * OpenAPI spec version: 0.1.0
 */
import type { DifficultyParamsAlgorithm } from "./difficultyParamsAlgorithm";

export interface DifficultyParams {
  algorithm?: DifficultyParamsAlgorithm;
  /**
   * @minimum 2
   * @maximum 20
   */
  crossValidationFolds?: number;
  useTextFeatures?: boolean;
  useEmbeddingFeatures?: boolean;
  embeddingModel?: string;
}